requirements.txt
pandas
numpy
torch
scikit-learn
matplotlib
torchvision
kaggle
jupyter
webdataset
pillow
scikit-learn
scikit-image
tensorboard
gdown
opencv-python
mediapipe